Added support for Windows to remove `dist` folder in unit tests.

+++ tests/Feature/ExportTest.php
@@ -17,7 +17,11 @@ class ExportTest extends BaseTestCase
     {
         parent::setUp();
 
-        exec('rm -r '.__DIR__.'/dist');
+        if (strtoupper(substr(PHP_OS, 0, 3)) === 'WIN') {
+            exec('del '.__DIR__.'\dist /q');
+        } else {
+            exec('rm -r '.__DIR__.'/dist');        
+        }
 
         Route::get('/', function () {
             return static::HOME_CONTENT;
